+++ src/routes/proposalRoute.tsx
@@ -5,6 +5,7 @@ import { useMemo } from "react";
 import { useAtomValue } from "jotai";
 import { web3walletAtom } from "@/store/web3wallet.store";
 import usePendingProposals from "@/hooks/usePendingProposals";
+import sessionProposals from "@/data/mocks/sessionProposals";
 
 export const proposalRoute = createRoute({
   getParentRoute: () => rootRoute,
@@ -14,10 +15,13 @@ export const proposalRoute = createRoute({
 
     const web3wallet = useAtomValue(web3walletAtom);
     const pendingProposals = usePendingProposals(web3wallet);
-    const proposal = useMemo(
-      () => pendingProposals.data.find((elem) => elem.id === Number(params.id)),
-      [params.id, pendingProposals.data]
-    );
+    // const proposal = sessionProposals.noSupport;
+    const proposal = sessionProposals.required;
+    // const proposal = sessionProposals.many;
+    // const proposal = useMemo(
+    //   () => pendingProposals.data.find((elem) => elem.id === Number(params.id)),
+    //   [params.id, pendingProposals.data]
+    // );
 
     if (!proposal) {
       return <Navigate to="/" search={(search) => search} />;
